requirements.txt
nats-py
fastapi
uvicorn[standard]
pydantic
python-dotenv
aiofiles
